add jest tests
