+++ __tests__/index.test.js
@@ -1,0 +1,7 @@
+const generateRandomUsername = require('../index');
+
+describe('index test suite', () => {
+    it('should export a single function', () => {
+        expect(generateRandomUsername).toEqual(expect.any(Function))
+    })
+})
