+++ package.json
@@ -8,8 +8,14 @@
     "url": "https://github.com/EvanPeterson1324/friendly-username-generator"
   },
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "jest"
   },
   "author": "Evan Peterson",
-  "license": "ISC"
+  "license": "ISC",
+  "devDependencies": {
+    "jest": "^27.0.5"
+  },
+  "jest": {
+    "verbose": true
+  }
 }
