+++ src/animals.js
